fix: move get_kpi_data_from_id function and  and change API Endpoint
Moved `get_kpi_data_from_id` from `kpi_view.py` to `kpi_controller.py`
and changed all imports of the same function in
`kpi_view.py`,`anomaly_data_view.py` and `config_setting_view.py`

changed the API Endpoint from `/config/<kpi_id>/multidim_status` to
`/config/dashboard_config`. Modified the API to take input parameters from
Args and modified the return object so facilitate accomodation of more
data.
