+++ chaos_genius/views/config_setting_view.py
@@ -2,12 +2,12 @@
 """Base views for config settings and onboarding."""
 from flask import Blueprint, current_app, request, jsonify
 
-import logging
 from chaos_genius.databases.models.data_source_model import DataSource
 from chaos_genius.databases.models.kpi_model import Kpi
 from chaos_genius.alerts.slack import trigger_overall_kpi_stats
 from chaos_genius.views.kpi_view import kpi_aggregation
 from chaos_genius.databases.models.config_setting_model import ConfigSetting
+from chaos_genius.controllers.kpi_controller import get_kpi_data_from_id
 from chaos_genius.controllers.config_controller import (
     get_modified_config_file,
     get_config_object,
@@ -17,27 +17,6 @@ from chaos_genius.controllers.config_controller import (
 from chaos_genius.databases.db_utils import chech_editable_field
 
 blueprint = Blueprint("config_settings", __name__)
-
-logger = logging.getLogger(__name__)
-
-def get_kpi_data_from_id(n: int) -> dict:
-    """Returns the corresponding KPI data for the given KPI ID
-    from KPI_DATA.
-
-    :param n: ID of KPI
-    :type n: int
-
-    :raises: ValueError
-
-    :returns: KPI data
-    :rtype: dict
-    """
-    # TODO: Move to utils module
-
-    kpi_info = Kpi.get_by_id(n)
-    if kpi_info and kpi_info.as_dict:
-        return kpi_info.as_dict
-    raise ValueError(f"KPI ID {n} not found in KPI_DATA")
 
 @blueprint.route("/onboarding-status", methods=["GET"])
 def get_onboarding_status():
@@ -194,19 +173,26 @@ def get_config_meta_data(config):
         )
         return jsonify({"message": str(err), "status": "failure"})
 
-@blueprint.route("/<int:kpi_id>/multidim_status", methods=["GET"])
-def multidim_status(kpi_id):
+@blueprint.route("/dashboard_config", methods=["GET"])
+def multidim_status():
     """check the number of dimensions in the kpi and return True if more than one dimension is present
     False otherwise"""
-    multidim_status = False
+    status, message = "", ""
+    data = {}
     try:
+        kpi_id = request.args.get("kpi_id")
+        multidim_status = False
         kpi_info = get_kpi_data_from_id(kpi_id)
         dimensions = kpi_info["dimensions"]
         if len(dimensions)>1:
             multidim_status = True
+        data["multidim_status"]=multidim_status
+        status="success"
     except Exception as err:
-        logger.info(f"Error Found: {err}")
-    return jsonify({"multidim_status":multidim_status, "msg": ""})
+        status = "failure"
+        current_app.logger.info(f"Error in fetching Dashboad Config Data: {err}")
+        message = str(err)
+    return jsonify({"data": data, "msg": message, "status": status})
 
 @blueprint.route("/update", methods=["PUT"])
 def edit_config_setting():
